+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/manager/ChangeRequestManager.java
@@ -245,7 +245,7 @@ public class ChangeRequestManager {
             throws SchedulerException, ValidationException {
         UserDTO jobUser = null;
         try {
-            jobUser = userManager.getById(AuthUtil.getCurrentUser().getId());
+            jobUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
         } catch (UserRetrievalException ex) {
             LOGGER.error("Could not find user to execute job.");
         }
