+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/DeveloperManager.java
@@ -376,7 +376,7 @@ public class DeveloperManager extends SecuredManager {
 
         UserDTO jobUser = null;
         try {
-            jobUser = userManager.getById(AuthUtil.getCurrentUser().getId());
+            jobUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
         } catch (UserRetrievalException ex) {
             LOGGER.error("Could not find user to execute job.");
         }
@@ -414,7 +414,7 @@ public class DeveloperManager extends SecuredManager {
 
         UserDTO jobUser = null;
         try {
-            jobUser = userManager.getById(AuthUtil.getCurrentUser().getId());
+            jobUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
         } catch (UserRetrievalException ex) {
             LOGGER.error("Could not find user to execute job.");
         }
