+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/PromotingInteroperabilityManager.java
@@ -64,7 +64,7 @@ public class PromotingInteroperabilityManager {
 
         UserDTO jobUser = null;
         try {
-            jobUser = userManager.getById(AuthUtil.getCurrentUser().getId());
+            jobUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
         } catch (UserRetrievalException ex) {
             LOGGER.error("Could not find user to execute job.");
         }
