+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/realworldtesting/manager/RealWorldTestingManager.java
@@ -80,7 +80,7 @@ public class RealWorldTestingManager {
         List<RealWorldTestingUpload> rwts = parseCsvFile(file);
         startRwtUploadJob(rwts);
 
-        UserDTO currentUser = userManager.getById(AuthUtil.getCurrentUser().getId());
+        UserDTO currentUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
         RealWorldTestingUploadResponse response = new RealWorldTestingUploadResponse();
         response.setEmail(currentUser.getEmail());
         response.setFileName(file.getName());
@@ -97,7 +97,7 @@ public class RealWorldTestingManager {
         rwtUploadJob.setGroup(SchedulerManager.CHPL_BACKGROUND_JOBS_KEY);
         JobDataMap jobDataMap = new JobDataMap();
         jobDataMap.put(RealWorldTestingUploadJob.RWT_UPLOAD_ITEMS, rwts);
-        jobDataMap.put(RealWorldTestingUploadJob.USER_KEY, userManager.getById(AuthUtil.getCurrentUser().getId()));
+        jobDataMap.put(RealWorldTestingUploadJob.USER_KEY, userManager.getById((Long) AuthUtil.getCurrentUser().getId()));
         rwtUploadJob.setJobDataMap(jobDataMap);
         rwtUploadTrigger.setJob(rwtUploadJob);
         rwtUploadTrigger.setRunDateMillis(System.currentTimeMillis() + DELAY_BEFORE_JOB_START);
