+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/surveillance/report/SurveillanceReportManager.java
@@ -186,7 +186,7 @@ public class SurveillanceReportManager extends SecuredManager {
             throws ValidationException, SchedulerException, UserRetrievalException {
         UserDTO jobUser = null;
         try {
-            jobUser = userManager.getById(AuthUtil.getCurrentUser().getId());
+            jobUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
         } catch (UserRetrievalException ex) {
             LOGGER.error("Could not find user to execute job.");
         }
@@ -527,7 +527,7 @@ public class SurveillanceReportManager extends SecuredManager {
             throws ValidationException, SchedulerException, UserRetrievalException {
         UserDTO jobUser = null;
         try {
-            jobUser = userManager.getById(AuthUtil.getCurrentUser().getId());
+            jobUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
         } catch (UserRetrievalException ex) {
             LOGGER.error("Could not find user to execute job.");
         }
